requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asttokens==2.4.1
attrs==24.2.0
autograd==1.7.0
autoray==0.7.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cirq==1.4.1
cirq-aqt==1.4.1
cirq-core==1.4.1
cirq-google==1.4.1
cirq-ionq==1.4.1
cirq-pasqal==1.4.1
cirq-rigetti==1.4.1
cirq-web==1.4.1
click==8.1.7
cloudpickle==3.0.0
cocotb==1.9.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
cudensitymat-cu11==0.0.5
cuquantum-cu11==24.11.0
custatevec-cu11==1.7.0
cutensor-cu11==2.0.2
cutensornet-cu11==2.6.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.9
duet==0.2.9
edalize==0.5.4
exceptiongroup==1.2.2
executing==2.1.0
find_libpython==0.4.0
fonttools==4.54.1
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.65.0
grpc-interceptor==0.15.4
grpcio==1.67.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
ibm-cloud-sdk-core==3.22.0
ibm-platform-services==0.59.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mitiq==0.41.0
mpmath==1.3.0
msgpack==1.1.0
mypy==1.13.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
packaging==23.2
pandas==2.2.3
parso==0.8.4
pbr==6.1.0
PennyLane==0.39.0
PennyLane_Lightning==0.39.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.9.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pygad==3.3.1
Pygments==2.18.0
PyJWT==2.9.0
pylatexenc==2.10
PyMatching==2.2.1
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyquil==4.14.3
pyspnego==0.11.1
python-dateutil==2.9.0.post0
python-rapidjson==1.20
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qcs-api-client-common==0.10.0
qcs-sdk-python==0.20.1
qiskit==1.1.1
qiskit-aer-gpu-cu11==0.15.1
qiskit-algorithms==0.3.1
qiskit-ibm-provider==0.11.0
qiskit-ibm-runtime==0.32.0
qiskit-research==0.0.4
quil==0.13.2
requests==2.32.3
requests_ntlm==1.3.0
rpcq==3.11.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
rustworkx==0.15.1
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
stevedore==5.3.0
stim==1.14.0
symengine==0.13.0
sympy==1.13.3
tabulate==0.9.0
testtools==2.7.2
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
types-Deprecated==1.2.9.20240311
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vcdvcd==2.3.6
verb==0.1.0
vunit_hdl==4.7.0
wasmtime==25.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
yowasp-runtime==1.65
yowasp-yosys==0.46.0.0.post790
zipp==3.20.2
